fix(dcd): pull apart ToolbarPlugin
