+++ src/components/lexical/plugins/ToolbarPlugin/constants.ts
@@ -1,0 +1,1 @@
+export const LOW_PRIORITY = 1
