+++ src/components/lexical/plugins/ToolbarPlugin/utils.ts
@@ -1,0 +1,18 @@
+export const getSelectedNode = (selection) => {
+  const anchor = selection.anchor
+  const focus = selection.focus
+  const anchorNode = selection.anchor.getNode()
+  const focusNode = selection.focus.getNode()
+
+  if (anchorNode === focusNode) {
+    return anchorNode
+  }
+
+  const isBackward = selection.isBackward()
+
+  if (isBackward) {
+    return $isAtNodeEnd(focus) ? anchorNode : focusNode
+  } else {
+    return $isAtNodeEnd(anchor) ? focusNode : anchorNode
+  }
+}
